backend/requirements.txt
asgiref==3.8.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
Django==5.1.7
django-rest-knox==5.0.2
djangorestframework==3.16.0
dynaconf==3.2.10
hvac==2.3.0
idna==3.10
loguru==0.7.3
MarkupSafe==3.0.2
pycparser==2.22
requests==2.32.3
shortuuid==1.0.13
sqlparse==0.5.3
typing_extensions==4.13.0
urllib3==2.3.0
validators==0.34.0
